Update main.py
Replaced spacy with nltk and cleaned up the code. Also added a logo to the first tab anchored to the bottom right as well as a message in the second tab and I am working on getting log contents displayed in the second tab even if no analysis is done.